Cast `isValid` property to integer for consistent serialization across form components.

+++ src/Component/Form/UICheckboxRadioComponent.php
@@ -155,7 +155,7 @@ final class UICheckboxRadioComponent extends AbstractComponent
             'field' => $this->name ?: $this->id,
             'value' => $this->value,
             'error' => $this->errorMessage,
-            'isValid' => empty($this->errorMessage),
+            'isValid' => (int) empty($this->errorMessage),
             'type' => $this->type
         ]);
     }
