+++ src/Component/Form/UIInputComponent.php
@@ -75,7 +75,7 @@ final class UIInputComponent extends AbstractComponent
             'field' => $this->name ?: $this->id,
             'value' => $this->value,
             'error' => $this->errorMessage,
-            'isValid' => empty($this->errorMessage)
+            'isValid' => (int) empty($this->errorMessage)
         ]);
     }
