+++ src/Component/Form/UIToggleComponent.php
@@ -93,7 +93,7 @@ final class UIToggleComponent extends AbstractComponent
             'field' => $this->name ?: $this->id,
             'value' => $this->value,
             'error' => $this->errorMessage,
-            'isValid' => empty($this->errorMessage),
+            'isValid' => (int) empty($this->errorMessage),
             'type' => 'toggle'
         ]);
     }
